Restructure project under oxeign package
