+++ .env.example
@@ -4,7 +4,7 @@ API_HASH=your_api_hash
 BOT_TOKEN=your_bot_token
 
 # Database
-MONGO_URI=mongodb://localhost:27017/master_guardian
+MONGO_URI=mongodb://localhost:27017/oxeign
 
 # Bot control
 OWNER_ID=123456789
