+++ oxeign/swagger/__init__.py
@@ -1,5 +1,5 @@
 from motor.motor_asyncio import AsyncIOMotorClient
-from config import MONGO_URI
+from oxeign.config import MONGO_URI
 
 client = AsyncIOMotorClient(MONGO_URI)
 db = client.get_default_database()
